[refator] ffmpeg 구조 개선 #227
