+++ backend/src/main/java/com/seoulJJ/hypertube/global/exception/ErrorCode.java
@@ -6,8 +6,6 @@ import lombok.Getter;
 @AllArgsConstructor
 @Getter
 public enum ErrorCode {
-    //common
-    NULL_POINT(500, "G100", "NULL POINT EXCEPTION"),
 
     //auth
     BAD_CREDENTIALS(401, "AU100", "BAD CREDENTIALS"),
@@ -25,7 +23,7 @@ public enum ErrorCode {
     USER_TEXT_COLOR_WRONG_TYPE(401, "UR403", "USER TEXT COLOR CODE IS WRONG"),
     USER_ALREADY_ATTENDANCE(409, "UR301", "USER ALREADY ATTENDANCE"),
 
-    /** Common **/
+    //common
     INTERNAL_SERVER_ERR(500, "CM001","INTERNAL SERVER ERROR"),
     NOT_FOUND(404, "CM002", "NOT FOUND"),
     BAD_REQUEST(400, "CM003", "BAD REQUEST"),
@@ -34,7 +32,10 @@ public enum ErrorCode {
     PAGE_NOT_FOUND(404, "CM006", "PAGE NOT FOUND"),
     VALID_FAILED(400, "CM007" , "Valid Test Failed."),
     BAD_ARGU(400, "ARGUMENT-ERR-400", "잘못된 argument 입니다."),
-    UNREADABLE_HTTP_MESSAGE(400, "CM008", "유효하지 않은 HTTP 메시지입니다.");
+    UNREADABLE_HTTP_MESSAGE(400, "CM008", "유효하지 않은 HTTP 메시지입니다."),
+    FFMPEG_ERR(500, "CM009", "FFMPEG ERR"),
+    NULL_POINT(500, "CM010", "NULL POINT EXCEPTION");
+
 
 
     private final int status;
