+++ backend/src/main/java/com/seoulJJ/hypertube/global/exception/custom/FFmpegException.java
@@ -1,0 +1,9 @@
+package com.seoulJJ.hypertube.global.exception.custom;
+
+import com.seoulJJ.hypertube.global.exception.ErrorCode;
+
+public class FFmpegException extends CustomRuntimeException{
+    public FFmpegException(String message, ErrorCode errorCode) {
+        super(message, errorCode);
+    }
+}
