+++ nginx/nginx.conf
@@ -29,8 +29,8 @@ http {
         ssl_protocols TLSv1.2 TLSv1.3;
 
         location / {
-            proxy_pass http://10.12.8.2:3000;
-            # proxy_pass http://frontend:3000;
+            # proxy_pass http://10.12.8.2:3000;
+            proxy_pass http://frontend:3000;
             proxy_set_header Host $host;
             proxy_set_header X-Real-IP $remote_addr;
             proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
